поменял место для медиа файлов, переставил команды автора в приложение админ
